requirements.txt
langchain
faiss-cpu
openai
unstructured
xmltodict
tiktoken
streamlit
streamlit-chat